Docker-files/requirements.txt
Flask==1.0.2
pandas==0.24.2
scikit-learn==0.22.1
joblib
Pillow
scikit-image
requests